move basic arithmetic to mathics.eval.arithmetic (#789)
This PR starts to move the context independent arithmetic to a separate
module. Interestingly, just by moving around a little piece of code, the
doctest time seems to be reduced in another 10 seconds in my machine, it
is, a 13%.
